Move Reset Startpage to PreferenceSection
