+++ graylog2-web-interface/src/components/users/UserEdit/ProfileSection.jsx
@@ -2,7 +2,6 @@
 import * as React from 'react';
 import { Formik, Form } from 'formik';
 
-import StoreProvider from 'injection/StoreProvider';
 import { Button, Col, Row } from 'components/graylog';
 import { ReadOnlyFormGroup } from 'components/common';
 import User from 'logic/users/User';
@@ -10,8 +9,6 @@ import SectionComponent from 'components/common/Section/SectionComponent';
 
 import FullNameFormGroup from '../UserCreate/FullNameFormGroup';
 import EmailFormGroup from '../UserCreate/EmailFormGroup';
-
-const StartpageStore = StoreProvider.getStore('Startpage');
 
 type Props = {
   user: User,
@@ -31,25 +28,8 @@ const ProfileSection = ({
     email,
   } = user;
 
-  const _resetStartpage = () => {
-    // eslint-disable-next-line no-alert
-    if (window.confirm(`You are about to reset the startpage from user ${user.fullName}. Please confirm.`)) {
-      StartpageStore.set(user.id);
-    }
-  };
-
-  const isStartpageSet = !!user.startpage?.type;
-  const resetTitle = isStartpageSet ? 'Reset startpage' : 'No startpage set';
-  const resetButton = (
-    <Button title={resetTitle}
-            disabled={!isStartpageSet}
-            onClick={_resetStartpage}>
-      Reset Startpage
-    </Button>
-  );
-
   return (
-    <SectionComponent title="Profile" headerActions={resetButton}>
+    <SectionComponent title="Profile">
       <Formik onSubmit={onSubmit}
               initialValues={{ email, full_name: fullName }}>
         {({ isSubmitting, isValid }) => (
